wip
use margin to fix Chrome incorrectly jumping; even 1px is enough, as long as it offsets the element a tiny bit, chrome will anchor correctly when the lazy loaded image comes in. The padding is removed because the margin gets rendered separate from padding in Chrome, but Firefox doesn't render the margin.